requirements.txt
django
south
django-relationships
django-tastypie
django-model-utils
PIL
sorl-thumbnail
boto
django-storages
django-allauth